send game result

+++ src/LevelCotroller.ts
@@ -25,6 +25,9 @@ export class LevelContoller {
     pistolKillHandeler: () => void;
 
     private gameStarted = false
+    serverSendHandler: (data) => void;
+
+    droneKills = 0
 
     constructor() {
         this.playerUI = new PlayerUI(this)
@@ -34,6 +37,7 @@ export class LevelContoller {
     private droneFactoryHandler(event: any) {
         log('LevelContoller', event.event)
         if (event.event == 'kill') {
+            this.droneKills += 1
             let addScore = 10
             if (event.weapon == 'sword') {
                 addScore = 25
@@ -43,13 +47,21 @@ export class LevelContoller {
             }
             this.score.increase(addScore)
         } else if (event.event == 'smashPlayer') {
-            this.health.decrease(event.drone.attack)
-            this.playerUI.damage()
-            if (this.health.read() <= 0) {
-                this.playerUI.kill()
-                this.gameOver = true
-                if (this.skyBox != null) {
-                    this.skyBox.setBounds(false)
+            if (!this.gameOver) {
+                this.health.decrease(event.drone.attack)
+                this.playerUI.damage()
+                if (this.health.read() <= 0) {
+                    this.playerUI.kill()
+                    this.gameOver = true
+                    if (this.skyBox != null) {
+                        this.skyBox.setBounds(false)
+                    }
+                    this.serverSendHandler({
+                        type: 'score',
+                        level: this.level.read(),
+                        score: this.score.read(),
+                        kills: this.droneKills
+                    })
                 }
             }
         }
@@ -97,6 +109,7 @@ export class LevelContoller {
         this.level.set(0)
         this.factory.Reset()
         this.gameOver = false
+        this.droneKills = 0
         this.nextLevel()
         if (this.skyBox != null) {
             this.skyBox.setBounds(true)
@@ -161,5 +174,10 @@ export class LevelSystem implements ISystem {
 
     setPistolKillHandeler(param: () => void) {
         this.levelController.pistolKillHandeler = param
+    }
+
+    setServerSendHandler(param: (data) => void) {
+        this.levelController.serverSendHandler = param
+
     }
 }
